+++ tests/tests-shortcodes-pro.php
@@ -2,7 +2,7 @@
 
 class Tests_Shortcode_Pro extends WP_UnitTestCase {
 
-	public function NO_test_oembed_thumbnail_and_title() {
+	public function test_oembed_thumbnail_and_title() {
 
 		$properties = arve_get_host_properties();
 
@@ -22,13 +22,13 @@ class Tests_Shortcode_Pro extends WP_UnitTestCase {
 					'mode' => 'lazyload',
 				);
 
-				$this->assertNotContains( 'Error', arve_shortcode_arve( $attr ) );
-
 				#if( isset( $props['auto_title'] ) && $props['auto_title'] ) {
 				if( isset( $test['oembed_title'] ) ) {
-					$this->assertContains( '<h5 itemprop="name" class="arve-title">' . $test['title'] . '</h5>', arve_shortcode_arve( $attr ) );
+					$this->assertNotContains( 'Error', arve_shortcode_arve( $attr ) );
+					$this->assertContains( '<h5 itemprop="name" class="arve-title">' . $test['oembed_title'] . '</h5>', arve_shortcode_arve( $attr ) );
 				}
 				if( isset( $test['oembed_img'] ) ) {
+					$this->assertNotContains( 'Error', arve_shortcode_arve( $attr ) );
 					$this->assertRegex( '#<img [^>]*src="' . $test['oembed_img'] . '#', arve_shortcode_arve( $attr ) );
 				}
 			}
@@ -36,6 +36,36 @@ class Tests_Shortcode_Pro extends WP_UnitTestCase {
 
 		add_filter( 'shortcode_atts_arve',    'arve_pro_filter_atts_img_src', 8 );
 		add_filter( 'shortcode_atts_arve',    'arve_pro_filter_atts_img_src_srcset', 9 );
+	}
+
+	public function test_api_calls() {
+
+		$properties = arve_get_host_properties();
+
+		foreach ( $properties as $provider => $props ) :
+
+			if ( empty( $values['tests'] ) ) {
+				continue;
+			}
+
+			foreach ( $values['tests'] as $key => $test ) {
+
+				$attr = array(
+					'url'  => $test['url'],
+					'mode' => 'lazyload',
+				);
+
+				#if( isset( $props['auto_title'] ) && $props['auto_title'] ) {
+				if( isset( $test['api_title'] ) ) {
+					$this->assertNotContains( 'Error', arve_shortcode_arve( $attr ) );
+					$this->assertContains( '<h5 itemprop="name" class="arve-title">' . $test['api_title'] . '</h5>', arve_shortcode_arve( $attr ) );
+				}
+				if( isset( $test['api_img'] ) ) {
+					$this->assertNotContains( 'Error', arve_shortcode_arve( $attr ) );
+					$this->assertRegex( '#<img [^>]*src="' . $test['api_img'] . '#', arve_shortcode_arve( $attr ) );
+				}
+			}
+		endforeach;
 	}
 
 	public function test_thumbnail_byattachment_and_url() {