+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 ## Changelog ##
 
-### 2016-10-26 - Pro Addon 3.2.8 ###
+### 2016-10-30 - 7.9.7 ###
+
+* Fix: Thumbnail not correctly applied to <video> tag for self hosted videos.
+* Improved: CSS for self hosted videos.
+* New: Wistia Support
+
+### 2016-10-27 - Pro Addon 3.2.8 ###
 
 * Improved: oembed error message
 * Improved: Skip srcset function for PHP 5.3 and lower
 
-### 2016-10-26 - 7.9.6 ###
+### 2016-10-27 - 7.9.6 ###
 
 * Fix: Thumbnail not correctly applied to <video> tag for self hosted videos.
 * Improved: CSS for self hosted videos.